+++ editor/editor-core/example-helpers/confluence-macros/MacroComponent.tsx
@@ -1,15 +1,22 @@
 import React from 'react';
+import Spinner from '@atlaskit/spinner';
 import {
   Parameters,
   ExtensionParams,
 } from '@atlaskit/editor-common/extensions';
-
 type Props = {
   node: ExtensionParams<Parameters>;
 };
-
 export default ({ node }: Props) => {
-  return (
+  const [isVisible, setVisible] = React.useState(false);
+  let timer = React.useRef<NodeJS.Timeout>();
+  React.useEffect(() => {
+    timer.current = setTimeout(() => setVisible(true), 1000);
+    return () => timer.current && clearTimeout(timer.current);
+  }, []);
+  return !isVisible ? (
+    <Spinner />
+  ) : (
     <div>
       <p>THIS IS A FAKE VIEW FOR EXTENSIONS</p>
       <table>